add link-to-web from arklib
